requirements.txt
setuptools==65.5.0
wheel==0.40.0
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.1.0
pandas==1.5.3
numpy==1.24.3
python-multipart==0.0.6